Conflicts:
	spring4-mvc-ajax-example-master/src/main/java/com/millesBornes/web/model/Player.java

+++ spring4-mvc-ajax-example-master/src/main/java/com/millesBornes/web/model/Player.java
@@ -6,7 +6,7 @@ public class Player {
 
 	//Attributes
 	private String pseudo;
-	private int distance;
+	private double distance;
 	private int hunger;
 	private int sleep;
 	private ArrayList<Card> list_card_player;
@@ -79,7 +79,7 @@ public class Player {
 	}
 
 
-	public int getDistance() {
+	public double getDistance() {
 		return distance;
 	}
